플레이어 접속 및 움직임 완료
초반 로딩 오류 있음

+++ KHJ/Server/Server/Item.h
@@ -11,7 +11,7 @@ public:
 
 	bool isExist;									//아이템 존재 여부
 
-	void ItemSet(float x, float y, float z);
+	void ItemSet(int x, int y, int z);
 	void ItemAvoid(bool m_exist);
 
 };
